+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Components/Objects/ExploreComponent.cs
@@ -219,6 +219,15 @@ public class ExploreComponent : GH_Component, IGH_VariableParameterComponent
     // supplies the name and which namespace each end lives in, so a relation added to the spec appears by itself.
     foreach (var type in graph.RelationTypes)
     {
+      // A relation pointing AT geometry comes out as the geometry itself - a centerline is a curve you plug into
+      // Curve components, not an id you look up. Only the relations receive leaves unbaked reach here at all
+      // (ArtefactGraphCache.IsSurfacedGeometryRel), and Describe could only render their targets as bare indices.
+      if (ArtefactGraphCache.IsGeometryNamespace(type.TargetNamespace))
+      {
+        Add(values, Humanise(type.Name), Decode(graph, type.Rel, objK));
+        continue;
+      }
+
       if (ArtefactGraphCache.IsObjectNamespace(type.SourceNamespace))
       {
         Add(values, Humanise(type.Name), Describe(graph.Targets(type.Rel, objK), type.TargetNamespace, bundle));
@@ -330,6 +339,36 @@ public class ExploreComponent : GH_Component, IGH_VariableParameterComponent
     bundle.Nodes.TryGetValue(k, out var node) && node.Name is { Length: > 0 } name ? name
     : bundle.ObjectAppIds.TryGetValue(k, out var appId) ? appId
     : null;
+
+  /// <summary>
+  /// The geometry a relation points at, decoded into the active document's units - the same decode path the receive
+  /// itself uses, so a centerline lands exactly on the object it belongs to.
+  /// </summary>
+  /// <remarks>
+  /// Failures are swallowed rather than surfaced: an unreadable blob means one empty branch, and Explore is a
+  /// read-only inspection component - failing the solve over it would be worse than showing nothing. The decoder
+  /// already absorbs SGEO problems into its warning list; the catch covers the raw-blob path, which no relation
+  /// surfaced today reaches but a future one could.
+  /// </remarks>
+  private static List<RG.GeometryBase> Decode(ArtefactGraph graph, byte rel, int objK)
+  {
+    var warnings = new List<string>();
+    var result = new List<RG.GeometryBase>();
+    foreach (int geomK in graph.Targets(rel, objK))
+    {
+      try
+      {
+        result.AddRange(
+          ArtefactGeometryDecoder.DecodeGeometryIndex(geomK, graph.Bundle, graph.Bundle.Units, null, warnings)
+        );
+      }
+      catch (Exception ex) when (!ex.IsFatal())
+      {
+        // one unreadable fragment, not the whole port
+      }
+    }
+    return result;
+  }
 
   /// <summary>
   /// Turns dense ids into something readable, using the namespace the catalog declared for that end. Object and node
