+++ Connectors/Rhino/Speckle.Connectors.GrasshopperShared/Operations/Receive/GrasshopperArtefactObjectBuilder.cs
@@ -1,5 +1,4 @@
 using System.Globalization;
-using Rhino;
 using Speckle.Connectors.Common.Operations;
 using Speckle.Connectors.GrasshopperShared.HostApp;
 using Speckle.Connectors.GrasshopperShared.Parameters;
@@ -16,7 +15,6 @@ using Speckle.Sdk.Pipelines;
 using Speckle.Sdk.Pipelines.Receive.Artifacts;
 using DataObject = Speckle.Objects.Data.DataObject;
 using RG = Rhino.Geometry;
-using SOG = Speckle.Objects.Geometry;
 using SpeckleRenderMaterial = Speckle.Objects.Other.RenderMaterial;
 
 namespace Speckle.Connectors.GrasshopperShared.Operations.Receive;
@@ -361,13 +359,16 @@ internal sealed class GrasshopperArtefactObjectBuilder
       var instNode = bundle.Nodes[e.Dst];
       int defNodeK = instNode.DefRef!.Value;
       var definition = definitions[defNodeK];
-      var xf = BuildTransform(instNode.Transform, instNode.Units is { Length: > 0 } u ? u : DocUnits());
+      var xf = BuildTransform(
+        instNode.Transform,
+        instNode.Units is { Length: > 0 } u ? u : ArtefactGeometryDecoder.DocUnits()
+      );
 
       var proxy = new InstanceProxy
       {
         definitionId = definition.ApplicationId!,
         transform = Matrix4x4.Identity,
-        units = DocUnits(),
+        units = ArtefactGeometryDecoder.DocUnits(),
         maxDepth = 0,
       };
       var instanceWrapper = new SpeckleBlockInstanceWrapper
@@ -419,7 +420,9 @@ internal sealed class GrasshopperArtefactObjectBuilder
     {
       foreach (var solidK in solidKs)
       {
-        foreach (var g in DecodeGeometryIndex(solidK, bundle, fallbackUnits, sourceType, warnings))
+        foreach (
+          var g in ArtefactGeometryDecoder.DecodeGeometryIndex(solidK, bundle, fallbackUnits, sourceType, warnings)
+        )
         {
           result.Add((solidK, g));
         }
@@ -429,7 +432,9 @@ internal sealed class GrasshopperArtefactObjectBuilder
     {
       foreach (var e in displayEdges.OrderBy(x => x.Ord))
       {
-        foreach (var g in DecodeGeometryIndex(e.Dst, bundle, fallbackUnits, sourceType, warnings))
+        foreach (
+          var g in ArtefactGeometryDecoder.DecodeGeometryIndex(e.Dst, bundle, fallbackUnits, sourceType, warnings)
+        )
         {
           result.Add((e.Dst, g));
         }
@@ -438,20 +443,10 @@ internal sealed class GrasshopperArtefactObjectBuilder
     return result;
   }
 
-  /// <summary>The owning object's source type, where the SGEO primitive alone is ambiguous - see <see cref="AsSourceType"/>.</summary>
+  /// <summary>The owning object's source type, where the SGEO primitive alone is ambiguous - see
+  /// <see cref="ArtefactGeometryDecoder"/>.</summary>
   private static string? ObjectSourceType(ArtefactBundle bundle, int objK) =>
     bundle.ObjectProperties(objK).GetString("type") is { Length: > 0 } s ? s : null;
-
-  // SGEO encodes a single point and a whole point cloud under the same Points primitive, so Decode can only ever hand
-  // back a Pointcloud - and a Speckle Point came back as a one-point cloud, a different type to everything downstream
-  // [ENG-9162]. The object's source type is the discriminator the blob lacks. Mirrors RhinoHostObjectArtefactBuilder,
-  // which keys on Rhino's own ObjectType; here the send stamps the Speckle type instead.
-  private static Base AsSourceType(Base decoded, string? sourceType) =>
-    sourceType is not null
-    && sourceType.EndsWith(".Point", StringComparison.Ordinal)
-    && decoded is SOG.Pointcloud { points.Count: 3 } cloud
-      ? new SOG.Point(cloud.points[0], cloud.points[1], cloud.points[2], cloud.units)
-      : decoded;
 
   // The send side stores "units" per object as an EAV property, falling back to the bundle's overall units when absent
   // (mirrors RhinoHostObjectArtefactBuilder.ObjectUnits).
@@ -653,7 +648,7 @@ internal sealed class GrasshopperArtefactObjectBuilder
             // DocUnits — otherwise a 3dm member isn't rescaled and a block sent from a metre model lands 1000x too
             // small in a millimetre GH doc (mirrors RhinoHostObjectArtefactBuilder.BuildDefinitions).
             // no owning object row for a definition member, so no source type to disambiguate points with
-            foreach (var g in DecodeGeometryIndex(geomK, bundle, bundle.Units, null, warnings))
+            foreach (var g in ArtefactGeometryDecoder.DecodeGeometryIndex(geomK, bundle, bundle.Units, null, warnings))
             {
               Base? converted;
               try
@@ -705,7 +700,7 @@ internal sealed class GrasshopperArtefactObjectBuilder
           {
             definitionId = childDef.ApplicationId!,
             transform = Matrix4x4.Identity,
-            units = DocUnits(),
+            units = ArtefactGeometryDecoder.DocUnits(),
             maxDepth = 0,
           };
           members.Add(
@@ -801,7 +796,7 @@ internal sealed class GrasshopperArtefactObjectBuilder
       d[0] = d[5] = d[10] = d[15] = 1.0;
     }
 
-    double scale = Units.GetConversionFactor(units, DocUnits());
+    double scale = Units.GetConversionFactor(units, ArtefactGeometryDecoder.DocUnits());
     var t = RG.Transform.Identity;
     t.M00 = d[0];
     t.M01 = d[1];
@@ -820,145 +815,6 @@ internal sealed class GrasshopperArtefactObjectBuilder
     t.M32 = d[14];
     t.M33 = d[15];
     return t;
-  }
-
-  // Decodes one geometry index to Rhino geometry, scaled from its source units to DocUnits (SGEO carries its own
-  // units; 3dm uses the caller-supplied fallback) — mirrors RhinoHostObjectArtefactBuilder.DecodeGeometryIndex.
-  // ConvertToSpeckle (below) stamps the *active document's* units onto the converted Base without rescaling the
-  // numeric values, so mesh/3dm geometry must already be in DocUnits by the time it gets there.
-  private static List<RG.GeometryBase> DecodeGeometryIndex(
-    int geomK,
-    ArtefactBundle bundle,
-    string fallbackUnits,
-    string? sourceType,
-    List<string> warnings
-  )
-  {
-    if (!bundle.Geometries.TryGetValue(geomK, out var g))
-    {
-      return new List<RG.GeometryBase>();
-    }
-    if (g.Type == RawEncodingFormats.RHINO_3DM)
-    {
-      var geoms = RawEncodingToHost.Convert3dm(g.Content);
-      ApplyUnits(geoms, fallbackUnits);
-      return geoms;
-    }
-    if (g.IsSgeo)
-    {
-      // Meshes take the fast hand-rolled path (no Base allocation), scaled here.
-      if (SgeoDecoder.TryDecodeMesh(g.Content, out var sm))
-      {
-        var list = new List<RG.GeometryBase> { BuildMesh(sm) };
-        ApplyUnits(list, sm.Units);
-        return list;
-      }
-      // Curves, points, and point clouds: decode to a Speckle geometry object and convert via the shared Rhino ToHost
-      // converter (ConvertSpeckleGeometry below), which already scales from the decoded Base's own `units` to
-      // DocUnits (SpeckleToHostGeometryBaseTopLevelConverter) — so no extra ApplyUnits here, unlike mesh/3dm above.
-      // An unsupported/undecodable primitive degrades to nothing (with a warning) rather than aborting the receive.
-      Base? decoded = null;
-      try
-      {
-        decoded = AsSourceType(SgeoDecoder.Decode(g.Content), sourceType);
-        var converted = ConvertSpeckleGeometry(decoded);
-        if (converted.Count == 0)
-        {
-          warnings.Add($"Geometry {geomK} ({decoded.speckle_type}) did not convert to any native geometry.");
-        }
-        return converted;
-      }
-      catch (Exception ex) when (!ex.IsFatal())
-      {
-        string stage = decoded is null ? "decode" : $"convert of {decoded.speckle_type}";
-        warnings.Add($"Geometry {geomK} ({g.Type}) failed to {stage}: {ex.Message}");
-      }
-    }
-    return new List<RG.GeometryBase>();
-  }
-
-  // Speckle geometry object (from SgeoDecoder.Decode) → Rhino geometry via the shared Rhino ToHost converter (the
-  // same one Rhino's own artefact receive uses, ENG-8781). Mirrors RhinoHostObjectArtefactBuilder.ConvertSpeckleGeometry.
-  private static List<RG.GeometryBase> ConvertSpeckleGeometry(Base decoded) =>
-    SpeckleConversionContext
-      .Current.ConvertToHost(decoded)
-      .Select(pair => pair.Item1)
-      .OfType<RG.GeometryBase>()
-      .ToList();
-
-  private static void ApplyUnits(List<RG.GeometryBase> geoms, string? units)
-  {
-    if (units is not { Length: > 0 } u)
-    {
-      return;
-    }
-    var docUnits = DocUnits();
-    if (string.Equals(u, docUnits, StringComparison.OrdinalIgnoreCase))
-    {
-      return;
-    }
-    var t = RG.Transform.Scale(RG.Point3d.Origin, Units.GetConversionFactor(u, docUnits));
-    foreach (var geom in geoms)
-    {
-      geom.Transform(t);
-    }
-  }
-
-  // The active Rhino/GH document's units — what ConvertToSpeckle's converters (e.g. MeshToSpeckleConverter) stamp
-  // as the converted Base's "units" without rescaling, so all decoded geometry must land here before conversion.
-  private static string DocUnits() => RhinoDoc.ActiveDoc?.ModelUnitSystem.ToSpeckleString() ?? Units.Meters;
-
-  // SGEO neutral mesh → Rhino mesh (Speckle count-prefixed face format; mirrors RhinoHostObjectArtefactBuilder).
-  private static RG.Mesh BuildMesh(SgeoMesh sm)
-  {
-    var mesh = new RG.Mesh();
-    var v = sm.Vertices;
-    for (int i = 0; i + 2 < v.Length; i += 3)
-    {
-      mesh.Vertices.Add(v[i], v[i + 1], v[i + 2]);
-    }
-
-    var f = sm.Faces;
-    int p = 0;
-    while (p < f.Length)
-    {
-      int n = f[p];
-      if (n < 3)
-      {
-        n += 3; // legacy 0 -> triangle, 1 -> quad
-      }
-      if (n == 3 && p + 3 < f.Length)
-      {
-        mesh.Faces.AddFace(f[p + 1], f[p + 2], f[p + 3]);
-      }
-      else if (n == 4 && p + 4 < f.Length)
-      {
-        mesh.Faces.AddFace(f[p + 1], f[p + 2], f[p + 3], f[p + 4]);
-      }
-      else if (n > 4 && p + n < f.Length)
-      {
-        for (int k = 1; k < n - 1; k++)
-        {
-          mesh.Faces.AddFace(f[p + 1], f[p + 1 + k], f[p + 2 + k]);
-        }
-      }
-      else
-      {
-        break;
-      }
-      p += n + 1;
-    }
-
-    if (sm.Colors.Length == mesh.Vertices.Count && sm.Colors.Length > 0)
-    {
-      foreach (var argb in sm.Colors)
-      {
-        mesh.VertexColors.Add(System.Drawing.Color.FromArgb(argb));
-      }
-    }
-    mesh.Normals.ComputeNormals();
-    mesh.Compact();
-    return mesh;
   }
 
   // Resolves (and creates once) the nested collection-wrapper chain for the given scene-view segments under the root.
